ex05_ml_prediction_service/pyproject.toml
[project]
name = "ex05-ml-prediction-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "joblib>=1.4.2",
    "minio>=7.2.7",
    "pandas>=2.0.3",
    "pyarrow>=17.0.0",
    "python-dotenv>=1.0.0",
    "scikit-learn>=1.3.2",
    "streamlit>=1.40.1",
    "watchdog>=4.0.2",
]

[dependency-groups]
dev = [
    "flake8>=5.0.4",
    "pyment>=0.3.3",
    "pytest>=8.3.5",
]

[tool.pytest.ini_options]
pythonpath = [
    ".", "src"
]
